+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### [0.1.0](https://github.com/nire0510/onscroll/releases/tag/v0.1.0)
+
+- setStyle & callFunction arguments are now position (position.top, position.left) which holds last scroll position
+and direction (direction.north, direction.west, direction.east, direction.south) which holds last scroll direction
+- User can limit actions by scrolling direction using the direction property
+
 ### [0.0.2](https://github.com/nire0510/onscroll/releases/tag/v0.0.2)
 
 - Add support in multiple class names